+++ compiler/src/main/io.katana.compiler/io/katana/compiler/ast/type/AstTypeArray.java
@@ -14,14 +14,12 @@
 
 package io.katana.compiler.ast.type;
 
-import java.math.BigInteger;
-
 public class AstTypeArray extends AstType
 {
-	public BigInteger length;
+	public long length;
 	public AstType type;
 
-	public AstTypeArray(BigInteger length, AstType type)
+	public AstTypeArray(long length, AstType type)
 	{
 		this.length = length;
 		this.type = type;
